Push startup/shutdown rate stats logic into separate functions
This provides a cleaner interface to the main program and ensures
that all initialization and cleanup logic is fired correctly. The
stats display thread is also now only started when not reading a
saved capture file. It didn't make as much sense to do a real-time
display while parsing a local file and it led to different results
depending on when the display thread fired.