Change rate statistics mode switch to -s
Now -s is all it takes to enter rps stats mode. -t is still used
to set the display interval, but it has a sane default. There
are situations when it doesn't make sense to have to enter a
display interval, plus this makes it easier to start seeing data.